Adds support for es6 named import, overall cleanup and fixes (#188)
Almost a complete rewrite. While cleaning up for named import I decided to write tests in a more manageable way. Turned out quite a few of the typings had problems with currying. Also now with union types updates in 0.28 (hence the version requirement update) I was able to be quite specific with types (as usual — flow keeps impressing me). Fantasy-land and lenses are still not there. I need to wrap my head around typing it properly as it was created with dynamic types in mind.
I think the typings file is actually easier to navigate so I hope it's now more convenient to  improve the typings or fix bugs.