Re-write validate-defs to use npm-specific libdef code
This is the first step in adding support for "namespaces" to flow-typed. The
concept is pretty straightforward: Each libdef sits in a namespace, and each
namespace has its own set of rules around how libdefs are named, versioned,
tested, and installed. The current set of npm libdefs we have will go under the
"npm" namespace and this namespace will remain the default.

Namespaces will allow flow-typed to begin supporting non-npm libdefs such as the
core built-ins that are currently only embedded in Flow itself, a "staging"
namespace for less-vetted libdefs, a namespace for libdefs of atom services,
etc.